+++ create_dev_link.sh
@@ -1,11 +1,7 @@
 #!/bin/bash
-BASEDIR=$(cd `dirname $0` && pwd)
-CONFIG_DIR=$( test -z $1 && echo "/media/sdf1/home/fh8123789/.config/deluge" || echo "$1")
-[ -d "$CONFIG_DIR/plugins" ] || echo "Config dir "$CONFIG_DIR" is either not a directory or is not a proper deluge config directory. Exiting"
-[ -d "$CONFIG_DIR/plugins" ] || exit 1
-cd $BASEDIR
-test -d $BASEDIR/temp || mkdir $BASEDIR/temp
-export PYTHONPATH=$BASEDIR/temp
-/usr/bin/python setup.py build develop --install-dir $BASEDIR/temp
-cp $BASEDIR/temp/*.egg-link $CONFIG_DIR/plugins
-rm -fr $BASEDIR/temp
+cd /media/sdf1/home/fh8123789/deluge/texttab
+mkdir temp
+export PYTHONPATH=./temp
+/usr/bin/python setup.py build develop --install-dir ./temp
+cp ./temp/textTab.egg-link /media/sdf1/home/fh8123789/.config/deluge/plugins
+rm -fr ./temp
