--- deluge/plugins/texttab/data/config.glade
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<!DOCTYPE glade-interface SYSTEM "glade-2.0.dtd">
-<!--Generated with glade3 3.4.5 on Fri Aug  8 23:34:44 2008 -->
-<glade-interface>
-  <widget class="GtkWindow" id="window1">
-    <child>
-      <widget class="GtkHBox" id="prefs_box">
-        <property name="visible">True</property>
-        <child>
-          <widget class="GtkLabel" id="label1">
-            <property name="visible">True</property>
-            <property name="label" translatable="yes">Full path of file to show:</property>
-          </widget>
-        </child>
-        <child>
-          <widget class="GtkEntry" id="txt_path1">
-            <property name="visible">True</property>
-            <property name="can_focus">True</property>
-          </widget>
-          <packing>
-            <property name="position">1</property>
-          </packing>
-        </child>
-      </widget>
-    </child>
-  </widget>
-</glade-interface>
