+++ textTab.egg-info/SOURCES.txt
@@ -1,17 +1,14 @@
 setup.py
-deluge/__init__.py
-deluge/plugins/__init__.py
-deluge/plugins/texttab/__init__.py
-deluge/plugins/texttab/common.py
-deluge/plugins/texttab/core.py
-deluge/plugins/texttab/gtkui.py
-deluge/plugins/texttab/webui.py
-deluge/plugins/texttab/data/config.glade
-deluge/plugins/texttab/data/text_tab.glade
-deluge/plugins/texttab/data/texttab.js
 textTab.egg-info/PKG-INFO
 textTab.egg-info/SOURCES.txt
 textTab.egg-info/dependency_links.txt
 textTab.egg-info/entry_points.txt
-textTab.egg-info/namespace_packages.txt
 textTab.egg-info/top_level.txt
+texttab/__init__.py
+texttab/common.py
+texttab/core.py
+texttab/gtkui.py
+texttab/webui.py
+texttab/data/config.glade
+texttab/data/text_tab.glade
+texttab/data/texttab.js
