--- textTab.egg-info/namespace_packages.txt
@@ -1,2 +1,0 @@
-deluge
-deluge.plugins
