+++ textTab.egg-info/top_level.txt
@@ -1,1 +1,1 @@
-deluge
+texttab
