+++ texttab/data/config.glade
@@ -1,0 +1,70 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<!DOCTYPE glade-interface SYSTEM "glade-2.0.dtd">
+<!--Generated with glade3 3.4.5 on Fri Aug  8 23:34:44 2008 -->
+<glade-interface>
+  <widget class="GtkWindow" id="window1">
+    <child>
+      <widget class="GtkHBox" id="prefs_box">
+        <property name="visible">True</property>
+        
+         <child>
+              <widget class="GtkVBox" id="label_box">
+                <property name="visible">True</property>
+                <property name="orientation">vertical</property>
+        <child>
+          <widget class="GtkLabel" id="label1">
+            <property name="visible">True</property>
+            <property name="label" translatable="no">Path of file to show, filename becomes tab name:</property>
+          </widget>
+        </child>
+        <child>
+          <widget class="GtkLabel" id="label114">
+            <property name="visible">True</property>
+            <property name="xalign">0</property>
+            <property name="label" translatable="no">Poll time, seconds:</property>
+          </widget>
+          <packing>
+            <property name="x_options">GTK_FILL</property>
+          </packing>
+        </child>
+      </widget>
+  <packing>
+    <property name="position">0</property>
+  </packing>    
+  </child>
+        <child>
+              <widget class="GtkVBox" id="vbox3">
+                <property name="visible">True</property>
+                <property name="orientation">vertical</property>
+                             <child>
+                                  <widget class="GtkEntry" id="txt_path1">
+                                    <property name="visible">True</property>
+                                    <property name="can_focus">True</property>
+                                  </widget>
+                                  <packing>
+                                    <property name="position">1</property>
+                                  </packing>
+                                </child>
+                            <child>
+                              <widget class="GtkSpinButton" id="spin_poll">
+                                <property name="visible">True</property>
+                                <property name="can_focus">True</property>
+                                <property name="invisible_char">&#x25CF;</property>
+                                <property name="xalign">1</property>
+                                <property name="adjustment">5 1 3600 1 10 0</property>
+                                <property name="numeric">True</property>
+                                <property name="update_policy">if-valid</property>
+                              </widget>
+                              <packing>
+                                <property name="left_attach">1</property>
+                                <property name="right_attach">2</property>
+                                <property name="x_options"></property>
+                              </packing>
+                            </child>
+                </widget>
+          <packing>
+            <property name="position">0</property>
+          </packing>    
+          </child>
+  </widget>
+</glade-interface>
