+++ texttab/gtkui.py
@@ -1,22 +1,45 @@
 #
-# -*- coding: utf-8 -*-#
-
-# Copyright (C) 2016 haydent <www.httech.com.au>
+# gtkui.py
+#
+# Copyright (C) 2009 haydent <www.httech.com.au>
 #
 # Basic plugin template created by:
 # Copyright (C) 2008 Martijn Voncken <mvoncken@gmail.com>
 # Copyright (C) 2007-2009 Andrew Resch <andrewresch@gmail.com>
 # Copyright (C) 2009 Damien Churchill <damoxc@gmail.com>
-# Copyright (C) 2010 Pedro Algarvio <pedro@algarvio.me>
 #
-# This file is part of textTab and is licensed under GNU General Public License 3.0, or later, with
-# the additional special exception to link portions of this program with the OpenSSL library.
-# See LICENSE for more details.
+# Deluge is free software.
+#
+# You may redistribute it and/or modify it under the terms of the
+# GNU General Public License, as published by the Free Software
+# Foundation; either version 3 of the License, or (at your option)
+# any later version.
+#
+# deluge is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+# See the GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with deluge.    If not, write to:
+# 	The Free Software Foundation, Inc.,
+# 	51 Franklin Street, Fifth Floor
+# 	Boston, MA  02110-1301, USA.
+#
+#    In addition, as a special exception, the copyright holders give
+#    permission to link the code of portions of this program with the OpenSSL
+#    library.
+#    You must obey the GNU General Public License in all respects for all of
+#    the code used other than OpenSSL. If you modify file(s) with this
+#    exception, you may extend this exception to your version of the file(s),
+#    but you are not obligated to do so. If you do not wish to do so, delete
+#    this exception statement from your version. If you delete this exception
+#    statement from all source files in the program, then also delete it here.
 #
 
 import gtk
-import logging
 
+from deluge.log import LOG as log
 from deluge.ui.client import client
 from deluge.plugins.pluginbase import GtkPluginBase
 import deluge.component as component
@@ -24,8 +47,6 @@ import deluge.common
 from twisted.internet.task import LoopingCall
 from deluge.ui.gtkui.torrentdetails import Tab
 from common import get_resource
-
-log = logging.getLogger(__name__)
 
 class GtkUI(GtkPluginBase):
     def enable(self):
@@ -36,12 +57,10 @@ class GtkUI(GtkPluginBase):
         component.get("PluginManager").register_hook("on_show_prefs", self.on_show_prefs)
         
         self._text_tab = TextTab()
-        #self._text_tab2 = TextTab()
-        component.get("TorrentDetails").add_tab(self._text_tab)        
-        #component.get("TorrentDetails").add_tab(self._text_tab2)
+        component.get("TorrentDetails").add_tab(self._text_tab)
 
         self.textTab_timer = LoopingCall(self._text_tab.update_tab)
-        self.textTab_timer.start(5)
+        self.textTab_timer.start(10)
 
     def disable(self):
         component.get("Preferences").remove_page("textTab")
@@ -56,6 +75,7 @@ class GtkUI(GtkPluginBase):
             "path1":self.glade.get_widget("txt_path1").get_text()
         }
         client.texttab.set_config(config)
+        self._text_tab.update_tab()
 
     def on_show_prefs(self):
         client.texttab.get_config().addCallback(self.cb_get_config)
@@ -73,6 +93,7 @@ class TextTab(Tab):
         self._name = "TextTab"
         self._child_widget = glade_tab.get_widget("text_tab")
         self._tab_label = glade_tab.get_widget("text_tab_label")
+        self._tab_label_label = glade_tab.get_widget("text_tab_label_label")  
 
         self.textview = glade_tab.get_widget("text_tab_textview") 
         self.textview.set_editable(False)
@@ -81,14 +102,18 @@ class TextTab(Tab):
     def update_tab(self):
         client.texttab.get_text().addCallback(self.cb_get_text)
         
-    def cb_get_text(self, text):        
-        if text:
+    def cb_get_text(self, (result, filename, text)):
+        if result == 0:
+            log.info("textTab, file not found")
+            self.textview.get_buffer().set_text("file not found")
+            self._tab_label_label.set_text(filename)
+        elif result == 1:
             log.info("textTab, file contents: \n%s" % text)
-            self.textview.get_buffer().set_text(text) # would be better to only request and add new lines maybe rather than all
+            self.textview.get_buffer().set_text(text)
+            self._tab_label_label.set_text(filename)
+        elif result == 2:
+            log.info("textTab, no change")
 
-        else:
-            log.info("textTab, file not found or empty")
-            self.textview.get_buffer().set_text("file not found or empty")
 
     def __dest(self, widget, response):        
         widget.destroy()
