+++ texttab/webui.py
@@ -1,5 +1,5 @@
 #
-# setup.py
+# webui.py
 #
 # Copyright (C) 2009 haydent <www.httech.com.au>
 #
@@ -37,37 +37,20 @@
 #    statement from all source files in the program, then also delete it here.
 #
 
-from setuptools import setup
+from deluge.log import LOG as log
+from deluge.ui.client import client
+from deluge import component
+from deluge.plugins.pluginbase import WebPluginBase
 
-__plugin_name__ = "textTab"
-__author__ = "haydent"
-__author_email__ = "www.httech.com.au"
-__version__ = "0.1"
-__url__ = ""
-__license__ = "GPLv3"
-__description__ = ""
-__long_description__ = """"""
-__pkg_data__ = {__plugin_name__.lower(): ["template/*", "data/*"]}
+from common import get_resource
 
-setup(
-    name=__plugin_name__,
-    version=__version__,
-    description=__description__,
-    author=__author__,
-    author_email=__author_email__,
-    url=__url__,
-    license=__license__,
-    long_description=__long_description__ if __long_description__ else __description__,
 
-    packages=[__plugin_name__.lower()],
-    package_data = __pkg_data__,
+class WebUI(WebPluginBase):
 
-    entry_points="""
-    [deluge.plugin.core]
-    %s = %s:CorePlugin
-    [deluge.plugin.gtkui]
-    %s = %s:GtkUIPlugin
-    [deluge.plugin.web]
-    %s = %s:WebUIPlugin
-    """ % ((__plugin_name__, __plugin_name__.lower())*3)
-)
+    scripts = [get_resource("texttab.js")]
+
+    def enable(self):
+        pass
+
+    def disable(self):
+        pass
